nova-act/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
amazon-nova-act>=0.1.0
